+++ backend/src/main/java/com/stampcrush/backend/api/manager/reward/ManagerRewardFindApiController.java
@@ -3,7 +3,7 @@ package com.stampcrush.backend.api.manager.reward;
 import com.stampcrush.backend.api.manager.reward.request.RewardFindRequest;
 import com.stampcrush.backend.api.manager.reward.response.RewardFindResponse;
 import com.stampcrush.backend.api.manager.reward.response.RewardsFindResponse;
-import com.stampcrush.backend.application.manager.reward.RewardService;
+import com.stampcrush.backend.application.manager.reward.ManagerRewardFindService;
 import com.stampcrush.backend.application.manager.reward.dto.RewardFindDto;
 import com.stampcrush.backend.application.manager.reward.dto.RewardFindResultDto;
 import com.stampcrush.backend.config.resolver.OwnerAuth;
@@ -19,7 +19,7 @@ import java.util.List;
 @RequestMapping("/api/admin/customers/{customerId}/rewards")
 public class ManagerRewardFindApiController {
 
-    private final RewardService rewardService;
+    private final ManagerRewardFindService managerRewardFindService;
 
     @GetMapping
     public ResponseEntity<RewardsFindResponse> findRewards(
@@ -28,7 +28,7 @@ public class ManagerRewardFindApiController {
             @ModelAttribute @Valid RewardFindRequest request
     ) {
         RewardFindDto rewardFindDto = new RewardFindDto(customerId, request.getCafeId(), request.getUsed());
-        List<RewardFindResultDto> rewardFindResultDtos = rewardService.findRewards(rewardFindDto);
+        List<RewardFindResultDto> rewardFindResultDtos = managerRewardFindService.findRewards(rewardFindDto);
         List<RewardFindResponse> rewardFindResponses = rewardFindResultDtos.stream()
                 .map(RewardFindResponse::new)
                 .toList();
